MAJOR moving repositories to the infra layer, creating an db adapter in the interface layer and updating everything
