+++ backend-api/src/app/application/services/address/getAddress.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { AddressRepository } from '../../../domain/repositories/address.repository';
+import { AddressRepository } from '../../../infrastructure/repositories/address.repository';
 import { Address } from '../../../domain/models/address.model';
 
 @Injectable()
