+++ backend-api/src/app/application/services/article/createArticle.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { CreateArticleDTO } from '../../dtos/article/createArticle.dto';
-import { ArticleRepository } from '../../../domain/repositories/article.repository';
+import { ArticleRepository } from '../../../infrastructure/repositories/article.repository';
 import { Article } from '../../../domain/models/article.model';
 
 @Injectable()
