+++ backend-api/src/app/application/services/article/deleteArticle.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { ArticleRepository } from '../../../domain/repositories/article.repository';
+import { ArticleRepository } from '../../../infrastructure/repositories/article.repository';
 import { GetArticleService } from './getArticle.service';
 import { GetUserService } from '../user/getUser.service';
 import { GetReviewService } from '../review/getReview.service';
