+++ backend-api/src/app/application/services/article/updateArticle.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { ArticleRepository } from '../../../domain/repositories/article.repository';
+import { ArticleRepository } from '../../../infrastructure/repositories/article.repository';
 import { Article } from '../../../domain/models/article.model';
 import { UpdateArticleDTO } from '../../dtos/article/updateArticle.dto';
 import { GetArticleService } from './getArticle.service';
