+++ backend-api/src/app/application/services/role/getRole.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { RoleRepository } from '../../../domain/repositories/role.repository';
+import { RoleRepository } from '../../../infrastructure/repositories/role.repository';
 import { Role } from '../../../domain/models/role.model';
 
 @Injectable()
