--- backend-api/src/app/domain/repositories/role.repository.ts
@@ -1,12 +1,0 @@
-import { CreateRoleDTO } from '../../application/dtos/role/createRole.dto';
-import { Role } from '@prisma/client';
-import { UpdateRoleDTO } from '../../application/dtos/role/updateRole.dto';
-
-export abstract class RoleRepository {
-  abstract findByName(name: string): Promise<Role | null>;
-  abstract findById(id: string): Promise<Role | null>;
-  abstract create(data: CreateRoleDTO): Promise<Role>;
-  abstract update(data: UpdateRoleDTO): Promise<Role>;
-  abstract delete(id: string): Promise<boolean>;
-  abstract findAll(): Promise<Role[]>;
-}
