--- backend-api/src/app/domain/repositories/user.repository.ts
@@ -1,29 +1,0 @@
-import { ArticleAuthor, User, UserRole } from '@prisma/client';
-import { CreateUserDTO } from '../../application/dtos/user/createUser.dto';
-import { UpdateUserDTO } from '../../application/dtos/user/updateUser.dto';
-
-export abstract class UserRepository {
-  abstract create(
-    data: CreateUserDTO,
-    homeAddressId: string,
-    jobAddressId: string
-  ): Promise<User>;
-
-  abstract update(
-    data: UpdateUserDTO,
-    homeAddressId: string | undefined,
-    jobAddressId: string | undefined,
-    roleChanged: boolean
-  ): Promise<User>;
-
-  abstract delete(id: string): Promise<boolean>;
-
-  abstract findByEmail(email: string): Promise<User | null>;
-  abstract findById(id: string): Promise<User | null>;
-  abstract findAll(): Promise<User[]>;
-  abstract findAllRoles(): Promise<UserRole[]>;
-  abstract findRolesByAuthorId(userId: string): Promise<UserRole[]>;
-  abstract findByAddressId(addressId: string): Promise<User[]>;
-  abstract findByArticleId(articleId: string): Promise<ArticleAuthor[]>;
-  abstract findByReviewId(reviewId: string): Promise<User[]>;
-}
