+++ backend-api/src/app/interface/adapter/database.adapter.ts
@@ -1,0 +1,7 @@
+export abstract class DatabaseAdapter<T> {
+  abstract findById(id: string): Promise<T | null>;
+  abstract findAll(): Promise<T[]>;
+  abstract create(data: Partial<T>): Promise<T>;
+  abstract update(data: Partial<T>): Promise<T>;
+  abstract delete(id: string): Promise<boolean>;
+}
